rework the rework of MWindowPropertyCache
While the idea and priciple are formidable i couldn't digest the
implementation.  Let's achieve the same without macros, pointer-to- members
and a dozen state variables.  This is accomplished by hooking the collector
methods directly up to QTimer::timeout().  Also track the validity of the
property values in the same hash as we track the request states.
Full details are in the class definition of MWindowPropertyCache.