fix decorated window stacking
...and make test15.py pass.  The problematic case was: map two undecorated
windows, unmap the topmost and map a decorated one.  Due to compareWindows()
failure the undecorated came to the top.

The stack before sorting was:
[ <decorator>, <undecorated>, <decorated> ]

and it became:
[ <decorated>, <decorator>, <undecorated> ]

So rather than swapping <decorator> and <undecorated> it brought the latter
on top.  This was because we told qStableSort() to order the decorator like
its managed window.  We expected that the order of <decorated> and <undecorated>
will be kept because they are not in relation, because of stable sorting.

However that could work only coincidally.  Since neither <decorator> and
<undecorated> are in relation the qStableSort() may (and in this case, did)
choose to keep their ordering, moving <undecorated> below.

 * src/mcompositemanager.cpp:
   -- compareDecorator(): decide the decorator's relation with the other
      window based on the relative position of its managed window.
      This will always be consistent.
   -- compareWindows(): add SORTING debugging facility