say goodbye to MRmiServer and MRmiClient
...and welcome the unified MRmi.

The incentive was to make MRmiClient keep the connection open rather than
dropping it every time it has sent something.  Then the hilarious level of
overengineering made me sick and all the hell broke lose.  Now we have a
single connection between the wm and the decorator which workth in both
directions and implemented by a single class.  Lots of repetitive code
has been removed and the new class preserves the intermediate structures
(the data stream and the output buffer) rather than reallocating them.
A long-standing bug has also been fixed in which the socket readers
(_q_readData() and _q_readyRead()) did not suck everything out of the
pipe, so requests could stuck there until a new one was sent.

  * decorators/libdecorator/mrmiserver.h
  * decorators/libdecorator/mrmiclient.cpp
  * decorators/libdecorator/mrmiclient.h
  * decorators/libdecorator/mrmiclient_p.h
  * decorators/libdecorator/mrmiserver.cpp
  * decorators/libdecorator/mrmiserver_p.h:
    Gone, replaced by mrmi.cpp and mrmi.h.
  * decorators/libdecorator/libdecorator.pro:
    Updated SOURCES and HEADERS.
  * decorators/libdecorator/mabstractappinterface.cpp:
    No need to #include mrmi.
  * decorators/libdecorator/mabstractdecorator.h
  * decorators/libdecorator/mabstractdecorator.cpp:
    No need for separate remote_compositor.
  * src/mdecoratorframe.h:
    Move instance() to the headers, so it can be inlined.
  * src/mdecoratorframe.cpp:
    Ensure that instanciating the class sets the singleton,
    so we don't need to rely on instance() to do that.
    Also export the parent (MCompositeManager) as soon as
    we connect to the decorator.
  * src/mcompositemanager.cpp:
    Just create the first and only MDecoratorFrame instance.