don't let ConfigureEvent:s interfere with animations
If you slow down the animation and map a decorated window with light
background you can see disgusting flickers.  This is because our beloved
decorator makes us reconfigure the window twice: once for portrait
decoration, then when it has managed to reorient itself, for landscape,
and we, as naive as we are, changed the position and the bounding box
of the MTexturePixmapItem immedeately.  This nonsense was earlier
hidden by the damage timer, but since we wait only one damage it has
become very visible.

The solution is to hold off the animation until the decorated window is
resized, just like we wait for damages.

 * src/mdecoratorframe.cpp (MDecoratorFrame::setManagedWindow):
   Unless we're told to @no_resize prepare the window for a likely resize.
 * src/mdecoratorframe.cpp (MDecoratorFrame::setDecoratorAvailableRect):
   Notify MCompositeManager about the resize immedeately, so it
   doesn't wait for the ConfigureEvent.
 * src/mcompositewindow.h:
   Merged damageReceived() with damageTimeout(), removed waitingForDamage().
 * src/mcompositewindow.cpp:
   * expectResize() added
   * damageReceived(): don't start q_fadein() until we got both the
     anticipated damages and resize, or until the timer has expired.
   * resize() added, makes the timer expire if all damages have arrived
 * src/mtexturepixmapitem_p.cpp (MTexturePixmapItem::resize):
   Chain back to MCompositeWindow::resize() so it will know about the resize.
 * src/mcompositemanager_p.h
 * src/mcompositemanager.h
 * src/mcompositemanager.cpp:
   * MCompositeManagerPrivate::damageEvent():
     don't check for waitingForDamage(), MCompositeWindow knows it
   * MCompositeManagerPrivate::configureEvent():
     Do not setPos() an animating window, unless the geometry has changed.
     (With expectResize() this is just a safety measure.)
   * MCompositeManagerPrivate::configureEvent():
     Accept a @nostacking parameter telling us not to bother with stacking.
   * MCompositeManager::expectResize(): fake a ConfigureEvent in order to
     have the item resized and positioned before the animation woould begin.
(cherry picked from commit a83d7dac05b32c9cf014c65ed7619dfe6b8cc2e4)