Goal: Simplify creating complex animation structures for windows
- Make sure that the precondition compositing is already started internally
when animation starts and disabled internally when it ends.
- Animation object  which is installed per window object
- Expose optional properties that can be overriden and chained to other
animation objects. This follows more closely the design and architecture of
Qt's animation framework (http://doc.trolltech.com/4.7/animation-
overview.html)
- Simpify beginAnimation/endAnimation. dont publicly expose atomic operations
beginAnimation / endAnimation. This should happen automatically behind the
scenes. beginAnimation() triggers when internal QPropertrtyAnimation for each
window gets started and endAnimation() triggeres when it ends.
- Dont manually handle zValue changes on animation start. Animation start
always ensure an item is visible.

Deprecated functions and classes:
- deprecate unflexible MCompwindowAnimator
- deprecate MCompositeManagerExtension virtual functions:
    virtual void windowShown();
    virtual void windowClosed();
    virtual void windowIconified();
    virtual void windowRestored();

New class
- MCompositeWindowAnimation which implements a default set of simple
animations for windows. For complex animations this can be subclassed and may
be extended using it's internal animation properties or add new ones by
hooking into it's QParallelAnimationGroup object
- One animation object per windows. To prevent random triggering of animations
objects, this is rare in the old code but does happen.

New signal in MCompositeManager:
- void windowBound(MCompositeWindow* window)

TODO:
- iconGeometry is not fixed in the un-pluginless compositor
- Centralize iconification  handling as well back to the compositor
- more cleanups in the visible signal

FIXES:
zValue based on types of windows instead of stacking list, as this does not
reflect the actual position if windows in the scene