fix infinitely looping RAISE_MATCHING
Bogus case: three windows are matching the criteria, two of them on the
top of the stack.  The lowest of them is raised, first_moved is set.
Then the (last-1)th is raised.  @next's index was @last but because of
(last-1) it takes its place.  Since @first_moved is already set it's
not set again.  So @next (last-1) is raised and it selects the original
(last-1), to be the @next.  QED.

  * src/mcompositemanager.cpp (RAISE_MATCHING):
    Rather than remembering _what_ we moved first, get its new place
    and then make sure we don't go beyond it.