Fixes: NB#264601 - Call log is not opened in foreground after tapping on missed call notification while there is a Messaging app with visible VKB opened in the background
Details:
When an application requestion the keyboard goes into the background we
have a race between the appearing window and the appearing keyboard.

Now the keyboard does not raise the transient parent anymore if it is
already mapped. Either it is already visible, or it is not supposed to
be visible.