Fixes: NB#278344 - A white vertical bar flickers in the left of the screen when creating an emergency call.
RevBy: Kimmo Hämälainen

Details: When emergency call is made call-ui is displayed in fullscreen window.
Fullscreen windows in LMT has statusbar hidden by default and only when
call is active status bar is shown. Also compositor itself displays decorator
for fullscreen windows without statusbar during call. Both compositor and LMT
monitors Phone.Call context property to know when call is stared.

Theoretically everything should be ok: emergency call is started, fullscreen
window is displayed, we are in call so we have statusbar visible; on compositor
side: window is fullscreen but we are in call, so we are checking that
statusbar geometry is set, so no decoration sould be displayed.

But because of bad timings fullscreen flag for window is set, compositor decides
that decoration is needed (because LMT didn't have time to react on call and
display statusbar) so decorator is displayed, then LMT shows statusbar and
statusbar geometry property is set, so mcompositor hides decorator. We see
flickering.

To fix this issue I just eliminated call-ui window from beeing decorated
during call.