be mercyful with other clients while animating
When an animation is started we grab the X server to prevent them rendering,
and we maintain the grab until the animation is over.  This can be problematic
if the swipe is synthetic, because the client cannot finish the motion, even
though we don't use the grab for anything, since we're waiting for motions.
To prevent this deadlock do release the grab after we haven't painted anything
for a while.

  * src/mcompositemanager.cpp (MSGrabber):
    Replaces MCompositeWindow::updateServerGrab() and friends.
    -- MCompositeWindow::we_have_grab		-> MSGrabber::has_grab
    -- MCompositeWindow::we_want_grab		-> MSGrabber::needs_grab
    -- MCompositeWindow::updateServerGrab()	-> MSGrabber::commit()
    -- MCompositeWindow::incomingCall()		-> MSGrabber::ungrab()
  * src/mcompositemanager.cpp (onFirstAnimationStarted, onAnimationsFinished):
    Do grabLater()/ungrab() here instead of MCompositeWindow::beginAnimation()
    and endAnimation().
  * src/mcompositemanager.cpp (sendSyntheticVisibilityEventsForOurBabies):
    Call it automatically on MSGrabber::ungrabbed().
  * src/mcompositemanager.cpp (MCompositeManager::runningInTestImage): bye-bye

				T  H  E     E  N  D