Fixes: NB#285253 - Window thumbnails in Recents do not make room for the swiped window
Details:
Make sure there are at least 150ms between ungrab and grab. This way
applications, especially the homescreen, can react to possible events
piled up during the grab.
If the user pauses a manual swipe in the middle we can end up in a
situation where a swipe animation runs without having a server grab. If
this turns out to be a problem we need another workaround.