Fixes: NB#287371 - "Settings" UI freezes after switching UI
Details:
Due to a bug somewhere else in the stack (maybe the service framework)
the _MEEGOTOUCH_WM_INVOKED_BY property sometimes points to a
non-existing window.
In this case the chaining animation has not been properly canceled.
This caused two problems
a) The window was transparent. The reason was that setting up an
animation changed the window to be transparent. This does not make any
sense these days and just hides bugs. This has been removed.
b) As the animation has never been canceled a crashing settings plugin
was leading to an infinite loop as compositor retried over and over
again to delete the window once the animation is done.
Now windows are no more transparent to avoid transparent windows in the
future and instead of canceling the animation altogether it is now
executed without animating the invoker window.