Add support for shaped windows in the renderer
Fix the last visible remaining issue in the renderer where shaped windows not rendered
properly. Evident in decorated window rendering garbage instead of the client.
In this case, only the decorator is rendered because the decorator’s shape changed and
the renderer thinks that the shape of the decorator remains unchanged. The visibility
algorithm thinks the client window is covered. To address this problem:

(1) The visibility algorithm is improved by changing how the bounding box is modified.
The bounding box is now updated by merging to itself rendered rectangles as long as
those rectangles share an edge or intersect with the existing bounding box.
(2) Support multi-region textures in geometry nodes by using a set of child geometry nodes
that differs in geometry from the main geometry node.
(3) Support shaped window events from MTexturePixmapItem and hook to the functionality
in MRender that adds regions in the GeometryNode of a window. This introduces the new
signal void MWindowPropertyCache::shapeUpdated().