Fixes: NB#296146 - startup time for applications increased by 400ms
Details:
As damage objects are created with XDamageReportNonEmpty there is a
race between compositor handling a damage event by calling
XDamageSubtract and the creation of damage events.
As we need exactly two damage events before we start the startup
animation we can run into a case where only one instead of two damage
events are created.
wo overcome this problem when mapping a window we always create the
damage object with XDamageReportRawRectangles. As soon as we are not
waiting for any more damage events before showing a window we switch to
XDamageReportNonEmpty (could be immediately after the next damage event
as some windows do not wait for damage).